show logical core percentage
